fix: change shard_size type from usize to u32

+++ tokio/src/runtime/time/entry.rs
@@ -672,7 +672,7 @@ cfg_rt! {
 }
 
 cfg_not_rt! {
-    fn generate_shard_id(shard_size: usize) -> u32 {
+    fn generate_shard_id(shard_size: u32) -> u32 {
         context::thread_rng_n(shard_size) % shard_size
     }
 }
